refactor: consolidate helper model resolution and fix systemPrompt passing
Extract shared helper model resolution (resolve name, create handler,
get client) into createHelperModelKit() to eliminate ~15 lines of
duplication across sessionDescription.ts, agentCreatorCommands.ts,
and textEnhancementUtils.ts.

Fix HIGH: pass systemPrompt to createResponse() in addition to
initializeMessages() so Anthropic/Google providers receive it
(they ignore the initializeMessages parameter).

Fix MEDIUM: collapse newlines in AI-generated descriptions to
prevent corrupting line-based ExecutionsTool output.
